Atualização

+++ GenerallySport/Controllers/PedidoVendaController.cs
@@ -93,7 +93,7 @@ namespace GenerallySport.Controllers
         public List<ItensPedidoVenda> GetItens()
         {
             PedidoVendaDAO itensPedidoVendaDAO = new PedidoVendaDAO();
-            var idCliente = long.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value);
+            var idCliente = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value);
             return itensPedidoVendaDAO.RetornarListaItensPedidoVenda(idCliente);
         }
 
@@ -103,7 +103,7 @@ namespace GenerallySport.Controllers
         public List<ItensPedidoVenda> GetItensPorId([FromRoute] int Id)
         {
             PedidoVendaDAO itensPedidoVendaDAO = new PedidoVendaDAO();
-            var idCliente = long.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value);
+            var idCliente = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value);
 
             return itensPedidoVendaDAO.RetornarListaItensPedidosVendaPorId(Id, idCliente);
